refactor: shared types 구조 정리
- shared 루트에 남아 있던 타입 파일을 역할별 파일로 분리
- 생성 요청/상태 관련 타입을 `types/generation.ts`로 이동
- 히스토리 사이드바 아이템 타입을 `types/historySidebar.ts`로 이동
- 타입 이동에 따른 import 경로 수정
- 미사용 `shared/types.ts` 파일 제거
