+++ src/shared/actions.ts
@@ -1,7 +1,7 @@
 'use server';
 
 import { privateFetch } from '@/shared/api/httpClient';
-import { GetGenerationStatusRequest, GetGenerationStatusResponse } from '@/shared/types';
+import { GetGenerationStatusRequest, GetGenerationStatusResponse } from '@/shared/types/generation';
 import { ApiResponse } from '@/shared/types/api';
 
 // AI 생성 상태 조회
