+++ src/shared/types/historySidebar.ts
@@ -1,0 +1,5 @@
+export interface HistorySidebarItem {
+  title: string;
+  date: string;
+  href: string;
+}
